pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cebus-mcp-scanner"
version = "0.1.0"
description = "Static security scanner for MCP servers and tools."
authors = [
  { name = "Cebus Security Labs" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.0",
  "pyyaml>=6.0.0"
]

[project.urls]
Homepage = "https://example.com/cebus-mcp-scanner"

[project.scripts]
mcp-scanner = "mcp_scanner.cli:run"

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.5.0",
  "tomli>=2.0.0; python_version < '3.11'"
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests"]

[tool.typer]
name = "mcp-scanner"

